% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlateRchs.R
\name{correlateRchs}
\alias{correlateRchs}
\title{Estimate Correlation Between Reaches}
\usage{
correlateRchs(
  redd_df = NULL,
  date_nm = "SurveyDate",
  redd_nm = "NewRedds",
  reach_nm = "Reach",
  use = "pairwise.complete.obs",
  ...
)
}
\arguments{
\item{redd_df}{dataframe containing redd data}

\item{date_nm}{quoted name of column in {redd_df} listing date of each survey}

\item{redd_nm}{quoted name of column in {redd_df} listing number of redds found during that survey}

\item{reach_nm}{quoted name of column in {redd_df} listing name of reach}
}
\value{
dataframe
}
\description{
Estimates correlation between reaches, to be used when calculating standard error across multiple reaches so reaches aren't necessarily treated independently
}
\author{
Kevin See
}
