% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_neterr.R
\name{predict_neterr}
\alias{predict_neterr}
\title{Predict Net Error}
\usage{
predict_neterr(data)
}
\arguments{
\item{data}{dataframe containing columns with covariates in observer error model}
}
\value{
tibble
}
\description{
Predict the net error of a steelhead redd survey, based on published model
}
\author{
Kevin See
}
