% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{removal_df}
\alias{removal_df}
\title{Number of Removals in Wenatchee}
\format{
A data frame with 36 rows and 5 variables:
\describe{
  \item{Year}{Spawn year}
  \item{Source}{Where the removal happened}
  \item{Origin}{Hatchery or natural}
  \item{rem}{How many fish were removed}
  \item{Area}{Which area in the DABOM model does this correspond to}
  ...
}
}
\source{
Michael Hughes at WDFW
}
\usage{
removal_df
}
\description{
A dataset containing how many fish were removed by origin, location and year (2014-2019)
}
\keyword{datasets}
