% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeRedds.R
\name{summarizeRedds}
\alias{summarizeRedds}
\title{Summarize Redds}
\usage{
summarizeRedds(
  redd_df = NULL,
  group_vars = c("River", "Reach", "Index", "SurveyType"),
  summ_vars = c("River", "Location"),
  new_redd_nm = "NewRedds",
  vis_redd_nm = "VisibleRedds",
  net_err_nm = "NetError",
  net_se_nm = "NetErrorSE",
  min_non0_wks = 3,
  min_redds = 2,
  use_cor = F,
  date_nm = "SurveyDate",
  redd_nm = "NewRedds",
  reach_nm = "Reach",
  use = "pairwise.complete.obs",
  method = c("pearson", "kendall", "spearman"),
  ...
)
}
\arguments{
\item{redd_df}{dataframe containing redd data}

\item{summ_vars}{vector of column names from {redd_df} to group summaries of redd results by}

\item{new_redd_nm}{quoted name of column in {redd_df} listing number of new redds found during that survey}

\item{vis_redd_nm}{quoted name of column in {redd_df} listing number of visible redds present during that survey}

\item{net_err_nm}{quoted name of column in {redd_df} listing estimate of net error for that survey}

\item{net_se_nm}{quoted name of column in {redd_df} listing standard error of net error estimate for that survey}

\item{min_non0_wks}{minimum number of weeks with at least one new redd observed}

\item{min_redds}{minimum number of total redds observed}

\item{use_cor}{should correlations between reaches be used in calculating standard errors? Default is {FALSE}}

\item{date_nm}{quoted name of column in {redd_df} listing date of each survey}

\item{redd_nm}{quoted name of column in {redd_df} listing number of redds found during that survey}

\item{reach_nm}{quoted name of column in {redd_df} listing name of reach}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}
}
\value{
dataframe
}
\description{
Estimates true number of redds across a series of reaches, using a Gaussian area-under-the-curve model where possible and incorporating estimates of observer error if possible. These estimates are then summarized across a user-defined spatial scale.
}
\author{
Kevin See
}
