% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thlwg_summ}
\alias{thlwg_summ}
\title{Mean Thalweg CV}
\format{
A data frame with 10 rows and 4 variables:
\describe{
  \item{Reach}{Reach label}
  \item{n_yrs}{Number of years measurements taken in that reach}
  \item{n_meas}{Number of distinct thalweg measurements taken in that reach}
  \item{MeanThalwegCV}{average CV of thalweg depth}
  ...
}
}
\source{
Michael Hughes at WDFW
}
\usage{
thlwg_summ
}
\description{
A dataset containing the mean thalweg CV of each reach in the Wenatchee, by using all measurements across all years
}
\keyword{datasets}
